requirements.txt
asgiref==3.5.2
attrs==22.1.0
black==22.8.0
certifi
click==8.1.3
Django==4.1.1
djangorestframework==3.14.0
iniconfig==1.1.1
isort==5.10.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.10.1
platformdirs==2.5.2
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
pytest==7.1.3
python-dotenv==0.21.0
pytz==2022.2.1
sqlparse==0.4.3
tomli==2.0.1
